# Support round-tripping nested values