… README.md
… csharp.test/TestFiles/map.parquet
… csharp.test/TestFiles/nested.parquet
… csharp.test/TestMaps.cs
… csharp/ColumnReader.cs
… csharp/LogicalBatchReader.cs
… csharp/LogicalBatchWriter.cs
… csharp/LogicalColumnReader.cs
… csharp/LogicalColumnStream.cs
… csharp/LogicalColumnWriter.cs
… csharp/LogicalStreamBuffers.cs
… csharp/LogicalWrite.cs
… csharp/Nested.cs
… csharp/Schema/SchemaUtils.cs
… csharp/TypeUtils.cs
… docs/Nested.md