This adds supports for writing values with a nested structure, ie. where there is a group node in the Parquet file schema that doesn't represent a list or map logical type. ~~This also introduces a breaking change when reading nested data, as now we will~~ When data is nested, there's now the option to read values wrapped in a new `Nested` type that is required to interpret the data correctly when there are multiple levels of nullability. Previously the actual level at which a value was null was ambiguous.

There's a lot more context in the comments on #221

This is based on @philjdf's work in https://github.com/philjdf/ParquetSharp/pull/4 but changed in a few ways:
* Generic classes are used to represent readers and writers for different element types rather than delegates that work with non-generic `Array` values, to hopefully make things more readable and debuggable, although this implementation is a fair bit more verbose
* Works with the performance improvement when reading nested array values from #295. Note that arrays of nested values will still use a list growing approach and need to read one value at a time.
* Read and write whole batches of required nested values in one operation
* When writing optional nested values, write runs of all null or all non-null values in one operation

There's also a breaking change in the `LogicalColumnStream` API where the protected `SchemaNodesPath` property has been replaced with `GetSchemaNodesPath`, but I don't expect many (probably any) users would be affected by this as I'm not aware of any use-cases for subclassing `LogicalColumnStream`.

Benchmark results vs the current master on my machine don't show any significant performance difference:
| Benchmark | Baseline | After |
| --- | ---: | ---: |
| DecimalRead | 82.18 +- 4.32 ms | 80.69 +- 1.323 ms |
| DecimalWrite | 112.3 +- 0.73 ms | 117.2 +- 0.68 ms |
| FloatTimeSeriesRead | 171.0 +- 3.29 ms | 161.5 +- 3.12 ms |
| FloatTimeSeriesWrite | 63.73 +- 0.503 ms | 65.62 +- 0.611 ms |
| FloatTimeSeriesWrite - chunked | 291.04 +- 4.219 ms | 291.9 +- 3.711 ms |
| FloatArrayTimeSeriesRead | 4.042 +- 0.0863 s | 3.95 +- 0.0581 s |