@adamreeve 2022-09-14
After some discussion with @marcin-krystianc, we thought it would be nice if we could introduce this change in a backwards compatible way so that users could continue to read nested data without needing the `Nested` wrapper type, so I've just pushed a change that does this. The `Nested` type still needs to be used if writing nested data, which wasn't possible before, but code that previously read nested data should now continue to work unchanged.

Users can opt in to using the `Nested` type by using it in the `TElement` parameter of the `LogicalColumnReader<TElement>` method if they have a need to accurately reproduce the nested structure of the data, or if using the non-generic `LogicalColumnReader` method there's a new override with a `useNesting` parameter.